// src/app/panchanga/model/day.ts
export class Day {
    date: Date;
    isAdikaMaasa: boolean;
    isKshayaMaasa: boolean;
    samvathsara: number;
    ayana: number;
    rutu: number;
    maasa: number;
    paksha: number;
    thithi: number[];
    vaasara: number;
    nakshathra: number;
    yoga: number;
    karana: number;
    shraddhaThithi: number[];
}